fix(shared,admin): address pool funding review feedback
