+++ packages/shared/src/__tests__/commitment-pooling-pool-funding-hooks.test.tsx
@@ -8,6 +8,7 @@ import { usePoolFunding } from "../hooks/commitment-pooling/usePoolFunding";
 import { createTestQueryClient, renderHookWithProviders } from "./test-utils";
 
 const GARDEN = "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa" as const;
+const OTHER_GARDEN = "0xbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb" as const;
 
 const mocks = vi.hoisted(() => ({
   getPoolFundingSnapshot: vi.fn(),
@@ -25,6 +26,7 @@ describe("usePoolFunding", () => {
       balance: { value: 25n, readAt: 2_000 },
       ledgerReadAt: 1_999,
       available: 20n,
+      fundingUnavailableReasons: [],
     });
   });
 
@@ -85,5 +87,44 @@ describe("usePoolFunding", () => {
     await waitFor(() => expect(result.current.hasStaleBalance).toBe(true));
     expect(result.current.snapshot?.balance?.value).toBe(25n);
     expect(result.current.lastReadAt).toBe(2_000);
+  });
+
+  it("does not reuse a prior pool's placeholder or retained balance", async () => {
+    let resolveOtherGarden: ((value: unknown) => void) | undefined;
+    mocks.getPoolFundingSnapshot
+      .mockResolvedValueOnce({
+        safe: GARDEN,
+        balance: { value: 25n, readAt: 2_000 },
+        ledgerReadAt: 1_999,
+        fundingUnavailableReasons: [],
+      })
+      .mockReturnValueOnce(
+        new Promise((resolve) => {
+          resolveOtherGarden = resolve;
+        })
+      );
+    const { result, rerender } = renderHookWithProviders(
+      ({ garden }: { garden: typeof GARDEN | typeof OTHER_GARDEN }) =>
+        usePoolFunding({ chainId: 42161, garden }),
+      {
+        initialProps: {
+          garden: GARDEN as typeof GARDEN | typeof OTHER_GARDEN,
+        },
+      }
+    );
+    await waitFor(() => expect(result.current.isSuccess).toBe(true));
+
+    rerender({ garden: OTHER_GARDEN });
+    expect(result.current.snapshot).toBeNull();
+
+    resolveOtherGarden?.({
+      safe: OTHER_GARDEN,
+      balance: null,
+      ledgerReadAt: 2_029,
+      fundingUnavailableReasons: ["balance_unreadable"],
+    });
+    await waitFor(() => expect(result.current.isSuccess).toBe(true));
+    expect(result.current.hasStaleBalance).toBe(false);
+    expect(result.current.snapshot?.balance).toBeNull();
   });
 });
