+++ packages/shared/src/hooks/commitment-pooling/usePoolFunding.ts
@@ -10,6 +10,8 @@ const FUNDING_REFRESH_INTERVAL_MS = 30_000;
 
 export function usePoolFunding(input: { chainId: number; garden: Address }) {
   const lastBalance = useRef<{
+    chainId: number;
+    garden: Address;
     safe: Address;
     balance: NonNullable<Awaited<ReturnType<typeof getPoolFundingSnapshot>>["balance"]>;
   } | null>(null);
@@ -19,16 +21,22 @@ export function usePoolFunding(input: { chainId: number; garden: Address }) {
     staleTime: STALE_TIME_MEDIUM,
     refetchInterval: FUNDING_REFRESH_INTERVAL_MS,
     refetchIntervalInBackground: false,
-    placeholderData: (previous) => previous,
   });
   const rawSnapshot = query.data ?? null;
   if (rawSnapshot?.safe && rawSnapshot.balance) {
-    lastBalance.current = { safe: rawSnapshot.safe, balance: rawSnapshot.balance };
+    lastBalance.current = {
+      chainId: input.chainId,
+      garden: input.garden,
+      safe: rawSnapshot.safe,
+      balance: rawSnapshot.balance,
+    };
   }
   const hasStaleBalance = Boolean(
     rawSnapshot?.safe &&
       !rawSnapshot.balance &&
       rawSnapshot.fundingUnavailableReasons.includes("balance_unreadable") &&
+      lastBalance.current?.chainId === input.chainId &&
+      lastBalance.current.garden.toLowerCase() === input.garden.toLowerCase() &&
       lastBalance.current?.safe.toLowerCase() === rawSnapshot.safe.toLowerCase()
   );
   const snapshot =
