+++ packages/shared/src/i18n/pt.json
@@ -4092,8 +4092,8 @@
   "cockpit.garden.pool.funding.feeBuffer": "Reserva conservadora para taxas",
   "cockpit.garden.pool.funding.fees.quote": "{fee} · {payer}",
   "cockpit.garden.pool.funding.fees.quoteUnavailable": "Cotação indisponível",
-  "cockpit.garden.pool.funding.fees.receiverPays": "paga o destinatário",
-  "cockpit.garden.pool.funding.fees.senderPays": "paga o remetente",
+  "cockpit.garden.pool.funding.fees.receiverPays": "paga pelo destinatário",
+  "cockpit.garden.pool.funding.fees.senderPays": "paga pelo remetente",
   "cockpit.garden.pool.funding.lastRead": "Última leitura",
   "cockpit.garden.pool.funding.limits.note": "Os limites restringem a execução da liquidação. Não são subtraídos do saldo do Safe.",
   "cockpit.garden.pool.funding.loading": "A carregar o financiamento do fundo comum",
