+++ packages/shared/src/modules/commitment-pooling/data-pool-funding.ts
@@ -1,10 +1,9 @@
 import type { PublicClient } from "viem";
 
-import { createPublicClientForChain } from "../../config/pimlico";
 import type { Address } from "../../types/domain";
-import { ERC20_BALANCE_ABI } from "../../utils/blockchain/abis/erc20";
 import { greenGoodsIndexer, type GraphQLReader } from "../data/graphql-client";
 import { address, integer, number, optionalInteger, type RawRow } from "./data-core";
+import { readPoolFundingChain } from "./data-pool-funding-chain";
 import {
   mapSettlementAccount,
   mapSettlementConfiguration,
@@ -16,165 +15,20 @@ import type {
   SettlementGardenRouteRecord,
 } from "./types-settlement";
 import {
-  calculateEffectiveZodiacAllowance,
   selectPoolFundingSnapshot,
-  type PoolFundingBalanceRead,
   type PoolFundingCalculationInput,
   type PoolFundingCommitment,
   type PoolFundingDeposit,
   type PoolFundingDisbursement,
   type PoolFundingExecution,
-  type PoolFundingFeeQuote,
   type PoolFundingPayoutPlan,
   type PoolFundingSnapshot,
-  type ZodiacAllowance,
 } from "./pool-funding";
 
 const PAGE_SIZE = 200;
 const LEDGER_MAX_AGE_SECONDS = 120;
 
-const BOOLEAN_PAUSED_ABI = [
-  {
-    type: "function",
-    name: "paused",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "bool" }],
-  },
-] as const;
-
-const GOOD_DOLLAR_ABI = [
-  ...ERC20_BALANCE_ABI,
-  ...BOOLEAN_PAUSED_ABI,
-  {
-    type: "function",
-    name: "getFees",
-    stateMutability: "view",
-    inputs: [
-      { name: "amount", type: "uint256" },
-      { name: "sender", type: "address" },
-      { name: "recipient", type: "address" },
-    ],
-    outputs: [
-      { name: "fee", type: "uint256" },
-      { name: "senderPays", type: "bool" },
-    ],
-  },
-] as const;
-
-const EXECUTOR_ABI = [
-  ...BOOLEAN_PAUSED_ABI,
-  {
-    type: "function",
-    name: "gardenRouteOf",
-    stateMutability: "view",
-    inputs: [{ name: "garden", type: "address" }],
-    outputs: [
-      {
-        name: "",
-        type: "tuple",
-        components: [
-          { name: "safe", type: "address" },
-          { name: "rolesModifier", type: "address" },
-          { name: "roleKey", type: "bytes32" },
-          { name: "allowanceKey", type: "bytes32" },
-          { name: "permissionsConfigHash", type: "bytes32" },
-          { name: "active", type: "bool" },
-        ],
-      },
-    ],
-  },
-  {
-    type: "function",
-    name: "gardenPeriodSpend",
-    stateMutability: "view",
-    inputs: [{ name: "garden", type: "address" }],
-    outputs: [
-      {
-        name: "",
-        type: "tuple",
-        components: [
-          { name: "periodStartedAt", type: "uint64" },
-          { name: "amount", type: "uint256" },
-        ],
-      },
-    ],
-  },
-  {
-    type: "function",
-    name: "maxTransferAmount",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint256" }],
-  },
-  {
-    type: "function",
-    name: "maxBatchAmount",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint256" }],
-  },
-  {
-    type: "function",
-    name: "maxBatchSize",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint16" }],
-  },
-  {
-    type: "function",
-    name: "maxFeeBps",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint16" }],
-  },
-  {
-    type: "function",
-    name: "maxFeeAmount",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint256" }],
-  },
-  {
-    type: "function",
-    name: "periodDuration",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint64" }],
-  },
-  {
-    type: "function",
-    name: "maxPeriodAmount",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint256" }],
-  },
-  {
-    type: "function",
-    name: "nativeFeeBalance",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [{ name: "", type: "uint256" }],
-  },
-] as const;
-
-const ROLES_ALLOWANCE_ABI = [
-  {
-    type: "function",
-    name: "allowances",
-    stateMutability: "view",
-    inputs: [{ name: "key", type: "bytes32" }],
-    outputs: [
-      { name: "refill", type: "uint128" },
-      { name: "maxRefill", type: "uint128" },
-      { name: "period", type: "uint64" },
-      { name: "timestamp", type: "uint64" },
-      { name: "balance", type: "uint128" },
-    ],
-  },
-] as const;
-
-interface PoolFundingLedger {
+export interface PoolFundingLedger {
   account: SettlementAccountRecord | null;
   route: SettlementGardenRouteRecord | null;
   sourceConfiguration: SettlementConfigurationRecord | null;
@@ -195,27 +49,6 @@ interface FundingPage {
   CommitmentFunding: RawRow[];
   Disbursement: RawRow[];
   SettlementExecution: RawRow[];
-}
-
-interface DirectReadResult {
-  balance: PoolFundingBalanceRead | null;
-  sourcePaused: boolean | null;
-  executorPaused: boolean | null;
-  tokenPaused: boolean | null;
-  liveRoute: {
-    safe: Address;
-    rolesModifier: Address;
-    allowanceKey: `0x${string}`;
-    active: boolean;
-  } | null;
-  feePolicy: { maxFeeBps: number; maxFeeAmount: bigint } | null;
-  feeQuotes: PoolFundingFeeQuote[];
-  rolesAllowanceRemaining: bigint | null;
-  periodAllowanceRemaining: bigint | null;
-  maxTransferAmount: bigint | null;
-  maxBatchAmount: bigint | null;
-  batchSizeLimit: number | null;
-  nativeFeeBalance: bigint | null;
 }
 
 function parseTimestamp(value: unknown): number | null {
@@ -332,7 +165,7 @@ async function queryPayoutRows(reader: GraphQLReader, ids: string[]): Promise<Ra
   return rows;
 }
 
-export async function getPoolFundingLedger(
+async function getPoolFundingLedger(
   sourceChainId: number,
   garden: Address,
   reader: GraphQLReader = greenGoodsIndexer,
@@ -546,326 +379,6 @@ export async function getPoolFundingLedger(
     caughtUpAt,
     readAt: now,
     coherent,
-  };
-}
-
-function routeTuple(value: unknown): DirectReadResult["liveRoute"] {
-  if (!value || (typeof value !== "object" && !Array.isArray(value))) return null;
-  const tuple = value as {
-    safe?: unknown;
-    rolesModifier?: unknown;
-    allowanceKey?: unknown;
-    active?: unknown;
-    0?: unknown;
-    1?: unknown;
-    3?: unknown;
-    5?: unknown;
-  };
-  const safe = address(tuple.safe ?? tuple[0]);
-  const rolesModifier = address(tuple.rolesModifier ?? tuple[1]);
-  const allowanceKey = tuple.allowanceKey ?? tuple[3];
-  if (!safe || !rolesModifier || typeof allowanceKey !== "string") return null;
-  return {
-    safe,
-    rolesModifier,
-    allowanceKey: allowanceKey.toLowerCase() as `0x${string}`,
-    active: (tuple.active ?? tuple[5]) === true,
-  };
-}
-
-function allowanceTuple(value: unknown): ZodiacAllowance | null {
-  if (!value || (!Array.isArray(value) && typeof value !== "object")) return null;
-  const row = value as {
-    0?: unknown;
-    1?: unknown;
-    2?: unknown;
-    3?: unknown;
-    4?: unknown;
-    refill?: unknown;
-    maxRefill?: unknown;
-    period?: unknown;
-    timestamp?: unknown;
-    balance?: unknown;
-  };
-  return {
-    refill: integer(row.refill ?? row[0]),
-    maxRefill: integer(row.maxRefill ?? row[1]),
-    period: integer(row.period ?? row[2]),
-    timestamp: integer(row.timestamp ?? row[3]),
-    balance: integer(row.balance ?? row[4]),
-  };
-}
-
-function periodSpendTuple(value: unknown): { periodStartedAt: bigint; amount: bigint } | null {
-  if (!value || (!Array.isArray(value) && typeof value !== "object")) return null;
-  const row = value as { 0?: unknown; 1?: unknown; periodStartedAt?: unknown; amount?: unknown };
-  return {
-    periodStartedAt: integer(row.periodStartedAt ?? row[0]),
-    amount: integer(row.amount ?? row[1]),
-  };
-}
-
-async function settledRead<T>(read: Promise<T>): Promise<T | null> {
-  try {
-    return await read;
-  } catch {
-    return null;
-  }
-}
-
-export async function readPoolFundingChain(
-  ledger: PoolFundingLedger,
-  garden: Address,
-  createClient: (chainId: number) => PublicClient = createPublicClientForChain as (
-    chainId: number
-  ) => PublicClient,
-  now = Math.floor(Date.now() / 1_000)
-): Promise<DirectReadResult> {
-  const source = ledger.sourceConfiguration;
-  const executor = ledger.executorConfiguration;
-  const safe = ledger.route?.safe ?? ledger.account?.account ?? null;
-  if (!source || !executor || !safe) {
-    return {
-      balance: null,
-      sourcePaused: null,
-      executorPaused: null,
-      tokenPaused: null,
-      liveRoute: null,
-      feePolicy: null,
-      feeQuotes: [],
-      rolesAllowanceRemaining: null,
-      periodAllowanceRemaining: null,
-      maxTransferAmount: null,
-      maxBatchAmount: null,
-      batchSizeLimit: null,
-      nativeFeeBalance: null,
-    };
-  }
-  const sourceClient = createClient(source.chainId);
-  const celoClient = createClient(executor.chainId);
-  const blockNumber = await settledRead(celoClient.getBlockNumber());
-  const block =
-    blockNumber === null ? null : await settledRead(celoClient.getBlock({ blockNumber }));
-  const [
-    balanceValue,
-    sourcePaused,
-    executorPaused,
-    tokenPaused,
-    routeValue,
-    feeBps,
-    feeAmount,
-    maxTransfer,
-    maxBatch,
-    maxBatchSize,
-    periodDuration,
-    maxPeriod,
-    periodSpendValue,
-    nativeFeeBalance,
-  ] = await Promise.all([
-    blockNumber === null
-      ? Promise.resolve(null)
-      : settledRead(
-          celoClient.readContract({
-            address: executor.gDollarToken,
-            abi: GOOD_DOLLAR_ABI,
-            functionName: "balanceOf",
-            args: [safe],
-            blockNumber,
-          })
-        ),
-    settledRead(
-      sourceClient.readContract({
-        address: source.localContract,
-        abi: BOOLEAN_PAUSED_ABI,
-        functionName: "paused",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "paused",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.gDollarToken,
-        abi: GOOD_DOLLAR_ABI,
-        functionName: "paused",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "gardenRouteOf",
-        args: [garden],
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "maxFeeBps",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "maxFeeAmount",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "maxTransferAmount",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "maxBatchAmount",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "maxBatchSize",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "periodDuration",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "maxPeriodAmount",
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "gardenPeriodSpend",
-        args: [garden],
-      })
-    ),
-    settledRead(
-      celoClient.readContract({
-        address: executor.localContract,
-        abi: EXECUTOR_ABI,
-        functionName: "nativeFeeBalance",
-      })
-    ),
-  ]);
-  const liveRoute = routeTuple(routeValue);
-  const allowanceValue = liveRoute
-    ? await settledRead(
-        celoClient.readContract({
-          address: liveRoute.rolesModifier,
-          abi: ROLES_ALLOWANCE_ABI,
-          functionName: "allowances",
-          args: [liveRoute.allowanceKey],
-        })
-      )
-    : null;
-  const allowance = allowanceTuple(allowanceValue);
-  const spend = periodSpendTuple(periodSpendValue);
-  const maxPeriodAmount = typeof maxPeriod === "bigint" ? maxPeriod : null;
-  const duration = typeof periodDuration === "bigint" ? periodDuration : null;
-  const spendExpired = Boolean(
-    spend && duration !== null && BigInt(now) >= spend.periodStartedAt + duration
-  );
-  const periodAllowanceRemaining =
-    maxPeriodAmount === null || !spend
-      ? null
-      : spendExpired
-        ? maxPeriodAmount
-        : maxPeriodAmount > spend.amount
-          ? maxPeriodAmount - spend.amount
-          : 0n;
-  const feePolicy =
-    (typeof feeBps === "number" || typeof feeBps === "bigint") && typeof feeAmount === "bigint"
-      ? { maxFeeBps: Number(feeBps), maxFeeAmount: feeAmount }
-      : null;
-
-  const openDisbursementIds = new Set(
-    ledger.disbursements
-      .filter(
-        (row) =>
-          row.source.toLowerCase() === safe.toLowerCase() &&
-          row.state !== "CONFIRMED" &&
-          row.state !== "CANCELLED"
-      )
-      .map((row) => row.disbursementId.toString())
-  );
-  const knownTransfers = [
-    ...ledger.disbursements
-      .filter((row) => openDisbursementIds.has(row.disbursementId.toString()))
-      .map((row) => ({ id: row.id, amount: row.amount, recipient: row.recipient })),
-    ...ledger.payoutPlans
-      .filter((plan) => plan.finalized)
-      .flatMap((plan) => plan.rows)
-      .filter((row) => row.disbursementId === null)
-      .map((row) => ({ id: row.id, amount: row.amount, recipient: row.recipient })),
-  ];
-  const feeQuotes = await Promise.all(
-    knownTransfers.map(async (row): Promise<PoolFundingFeeQuote> => {
-      const value = await settledRead(
-        celoClient.readContract({
-          address: executor.gDollarToken,
-          abi: GOOD_DOLLAR_ABI,
-          functionName: "getFees",
-          args: [row.amount, safe, row.recipient],
-        })
-      );
-      const tuple = value as readonly [bigint, boolean] | null;
-      return {
-        id: row.id,
-        amount: row.amount,
-        fee: tuple ? tuple[0] : null,
-        senderPays: tuple ? tuple[1] : null,
-        recipient: row.recipient,
-      };
-    })
-  );
-
-  return {
-    balance:
-      typeof balanceValue === "bigint" && blockNumber !== null && block
-        ? {
-            value: balanceValue,
-            blockNumber,
-            blockTimestamp: Number(block.timestamp),
-            readAt: now,
-          }
-        : null,
-    sourcePaused: typeof sourcePaused === "boolean" ? sourcePaused : null,
-    executorPaused: typeof executorPaused === "boolean" ? executorPaused : null,
-    tokenPaused: typeof tokenPaused === "boolean" ? tokenPaused : null,
-    liveRoute,
-    feePolicy,
-    feeQuotes,
-    rolesAllowanceRemaining: allowance
-      ? calculateEffectiveZodiacAllowance(allowance, BigInt(now))
-      : null,
-    periodAllowanceRemaining,
-    maxTransferAmount: typeof maxTransfer === "bigint" ? maxTransfer : null,
-    maxBatchAmount: typeof maxBatch === "bigint" ? maxBatch : null,
-    batchSizeLimit:
-      typeof maxBatchSize === "number" || typeof maxBatchSize === "bigint"
-        ? Number(maxBatchSize)
-        : null,
-    nativeFeeBalance: typeof nativeFeeBalance === "bigint" ? nativeFeeBalance : null,
   };
 }
